requirements.txt
numpy
matplotlib
tqdm 
libtiff
scipy
Pillow
scikit-image
opencv-python
imgaug
IPython
geopandas
keras
imantics
simplification
scikit-learn
chainer
tensorboard
tensorboardX
